backend/requirements.txt
fastapi
uvicorn
langgraph
langchain-groq
pydantic
sqlalchemy
pymysql
python-dotenv
